zabbix-mcp-server/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zabbix-mcp-server"
version = "0.1.0"
description = "Zabbix MCP Server — Model Context Protocol integration for Zabbix monitoring"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=24.0.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "respx>=0.21.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "bandit>=1.7.0",
    "types-python-dateutil>=2.9.0",
]

[project.scripts]
zabbix-mcp = "zabbix_mcp.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=zabbix_mcp --cov-report=term-missing"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"
ignore_missing_imports = true

[tool.coverage.run]
source = ["zabbix_mcp"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80
